getting post list in mypage with dummy done

+++ app/src/main/java/kr/co/youngcha/postup/Model/User/UserInfoResponseModel.kt
@@ -10,5 +10,5 @@ class UserInfoResponseModel (
     val nickname:String,
 
     @SerializedName("pu_post_ids")
-    val postIdList : List<Int>
+    val postIdList : List<Long>
 )
